ui/requirements.txt
streamlit
requests
pillow
tensorflow